pyproject.toml
[project]
name = "deliprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
    "pyusb>=1.3.1",
    "snoop>=0.6.0",
    "trio>=0.30.0",
]
